[FIX] sale_timesheet: Set the account_id on the analytic entries from timesheets
When we create a timesheet activity, it create an analytic line that allows to re-invoice the costs from the employee. If the account_id is not set on the analytic account line, the employee cost is not passed correctly on the Analytic Entries, which is not the behavior we should expect.

This commit simply adds a account_id on the employee and this account is returned with the cost and the uom by the method `_get_timesheet_cost`