[FIX] web_editor: stop resetting bg image of parent snippet on color change
Commit b83aa48d0d43511d000a4362b017e45e0f28fdb5 tried to solve the
problem but only solved a part of it. Indeed, if the snippet child
has a bg color option but not a bg image option, the event propagation
stops in the parent handler, when it is too late.

Solution is checking that the event comes from the same element on
which the callback is attached to.