[FIX] product: force empty context in product_template_action_all
This revision removes the context from the action:
7209f4718e8407fbb256dbc07c65068ab9af9e73

The thing is, in saas-11, the `product.template.type`
no longer has in its selection values `product`,
in the module `product`,
there is only `consu` and `service`.
the type `product` is added in the module `stock`

Therefore, this `product` cannot be used as default value
when only the module `product` is installed and not
`stock`.

We force the empty context for migration purposes,
so bases in ealier releases than saas-11 do not have
`'default_type': 'product'` within this action
when they are upgraded to saas-11.

opw-687986