[FIX] calendar: event display time timezone
If the method "_get_display_time" is called on an event and if the
user or partner has no timezone set. The string returned by the function
contains the value 'False' instead of a timezone.

To fix this, we replaced the "False" value by "UTC", because UTC is the
default timezone in Odoo.