[FIX] tools: prevent html_translate crash with malformed HTML
If an user stores malformed HTML code in a *translatable HTML field*,
the ORM won't prevent it (this could be a fix for master as it breaks
existing behaviours) and it could be rendered anyway by browsers.

When reading/prefetching the record, the translation function was
crashing because malformed HTML is obviously malformed XML (the
translation function uses XML internals).

This fix will return the source value as a fallback if the HTML field is
malformed (and thus not translatable) + log the error.