[REM] portal_sale: remove weird override of mail_mail processing
This code seems to add recipients of an email to the sale order followers.
However currently the custome of a sale order is already automatically
added in the followers. Adding other recipients should not be
something completely custom in sale order, moreover by overriding
mail_mail methods.

People should add followers when necessary and should avoid adding
too many people automatically. When you send an email to other people
than the customer those will be added as followers when replying.