Revert "[IMP] stock: Remove domain on source/destination on Operation Details"
This reverts commit 451f6310d1cda1424b6574b137696a6b847551f6.

The domain makes clear that the location to be chosen in is among the children of the location zones chosen in the picking.  That way, if the stock would be calculated in the past (e.g. also for the stock valuation report), the system is going to think it is maybe still internal, while it went out, ..., so those domains are crucial.