[FIX] website_blog: keep consistent cover classes when changing cover image
! Note forward port: this is a partial backport, ignore when conflict

For some strange reason, the cover_full class was added when the cover
image was changed, without checking if the cover_narrow class was
already there. The cover was then a hybrid cover_narrow cover_full.

The problem was reinforced by the graphene theme which does some
animation on cover_full covers.