[FIX] calendar: don't crash on missing optional dependency
For historical reasons, `vobject`* Python library is not mandatory in
Odoo.

Since refactoring 8151c29dea9e247d6e8f7749d4901ca911a90884,
get_ics_file() returns now a dictionary whose keys are calendar.event
ids and values are the corresponding ics file.
If `vobject` is not available, it returns an empty dictionary that
could cause calendar.py to crash because it always expects to find the
event's key.

Before this refactoring, this method was called and architectured in a
different manner, so no crash was happening.

This commit prevents the crash if `vobject` is not installed and logs a
warning.

*vobject allows to generate iCalendar files