Revert "[FIX] mail: onchange_template_id"
This reverts commit f118675d4e5adde657beda4d5bc48ddd842c8677.

This commit caused a side-effect that is worse than the issue
it tried to fix, which was purely cosmetic.
Now in mass-mail mode the template is only rendered once
and all recipients receive the same PDF, which completely
defeats the goal of a dynamic report. We have static
attachments for that.

Original fix related to opw-657152
Reverts PR #10568