[FIX] website_sale: recompute fiscal position
Following commit ace35e38f00, the shipping address can be changed after
the call to `checkout_form_save`. This is the case, for example, when
the user has the same shipping and invoicing address. We need to make
sure to recompute the fiscal position after the shipping address is
finally set.

opw-689281