[FIX] purchase_double_validation: from waiting approval to cancel
It wasn't possible to cancel a purchase order
within the state "Waiting for Approval", despite
the fact the button cancel is displayed.

Clicking on the cancel button just did nothing.

This is now properly working, thanks to the additional
path added by this revision, between the workflow states
`CheckForApproval` and `cancel`.

opw-688685