[FIX] website: fix FAQ Roller opening on old apple devices
On old apple devices (IPad 2, IPhone 4, ...), the FAQ Roller could not
be opened on click. This is because the snippet does not use standard
bootstrap correctly and put the toggle on a <div/> element instead of
a <a/> element.

This fix forces an empty click handler on these <div/>, which allows
standard bootstrap to work.