[FIX] product: fix duplicated creation variants
Move the code to manage the case of 'adding an attribute with only one value
should not recreate product' before to select the existing variants.

Before this commit,
When you create 1 new product.template, it create 1 variant.
When you add an attribute with only one value, it update the existing and
create a duplicate variant.

After this commit,
Only the existing variant is updated, and there are no more duplicated variant.

The fix:
We update the variant if only one value before to select the existing variant.
Thus, the updated variant is compared with the matrix and not duplicated.

This commit closes #13757