[FIX] payment_stripe: preventDefault
When a user uses Strip and clicks on "Pay & Confirm", a "Bad Request"
error can pop-up. This is because it tries to reload the page, while it
should only display the Strip pop-up.

This is simply due to a `preventDefault` done after the event is
performed, and not before.

opw-692341